requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.5
mysqlclient==2.0.3